Move library to root of repository
A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.

After this change, the installation scripts are no longer needed. These scripts would only work with the default sketchbook location.